unzip: move the Unzip method to the Downloader
In order to have the downloader manage all operations on the
downloaded file, the Unzip method has been added to it, thus making the
management/mocking easier.
