+++ billing/downloader_test.go
@@ -81,6 +81,58 @@ func TestBillingDownloader_Download(t *testing.T) {
 	}
 }
 
+func TestBillingDownloader_Unzip(t *testing.T) {
+	const (
+		zipNoSuchFile = "./test/no-such-file.zip"
+		zipDirPath    = "./test/test-dir.zip"
+		zipFilePath   = "./test/test.txt.zip"
+		zipFile       = "test.txt"
+	)
+	var tempDir string = os.TempDir() + "/"
+
+	t.Run("unzip directories and file", func(t *testing.T) {
+		var expectedPath string = tempDir
+		var expectedError error = nil
+
+		d := &billingDownloader{}
+		path, err := d.Unzip(zipDirPath, tempDir)
+		if path != expectedPath {
+			t.Errorf("Unzip invalid returned path: received=%q | expected=%q", path, expectedPath)
+		}
+		if err != expectedError {
+			t.Errorf("Unzip invalid error returned: received=%+v | expected=%+v", err, expectedError)
+		}
+	})
+
+	t.Run("unzip simple file", func(t *testing.T) {
+		var expectedPath string = tempDir + zipFile
+		var expectedError error = nil
+
+		d := &billingDownloader{}
+		path, err := d.Unzip(zipFilePath, tempDir)
+		if path != expectedPath {
+			t.Errorf("Unzip invalid returned path: received=%q | expected=%q", path, expectedPath)
+		}
+		if err != expectedError {
+			t.Errorf("Unzip invalid error returned: received=%+v | expected=%+v", err, expectedError)
+		}
+	})
+
+	t.Run("unzip no such file", func(t *testing.T) {
+		var expectedPath string = ""
+		var expectedError error = errors.New("open ./test/no-such-file.zip: no such file or directory")
+
+		d := &billingDownloader{}
+		path, err := d.Unzip(zipNoSuchFile, tempDir)
+		if path != expectedPath {
+			t.Errorf("Unzip invalid returned path: received=%q | expected=%q", path, expectedPath)
+		}
+		if err != nil && expectedError.Error() != err.Error() {
+			t.Errorf("Unzip invalid error returned: received=%+v | expected=%+v", err, expectedError)
+		}
+	})
+}
+
 func TestBillingDownloader_getAndCreateOutputPath(t *testing.T) {
 	const (
 		tempDownloadDir string = "/billingDownloader/"
